Harmonize DisplayName for argument values on Fixture and Method
